patch:  mode change check
Now throws an error when the mode that is tried to change to is invalid. Corresponding test had to be modified to have everything inside of the `with self.assertRaises(Exception)` statement. This is not patching the test to make the code work, this is how it should have been in the beginning I just messed up.